requirements.txt
kafka-python>=2.3.0

python-dotenv>=0.19.0